pyproject.toml
[project]
name = "llm-contract-analyzer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-genai>=1.29.0",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
